+++ main/module/vsh.py
@@ -1,7 +1,10 @@
 #!/usr/local/bin/python3
-import setting
-import libvirt, sys, sqlite3, subprocess, os
-import vsql, vansible, vhelp, vvirt, virty
+import libvirt
+import sys
+import sqlite3
+import subprocess
+import os
+from module import vsql, vansible, vhelp, vvirt, virty, setting
 
 
 #Class
@@ -9,16 +12,9 @@ class Color():
     BLACK     = '\033[30m'
     RED       = '\033[31m'
     GREEN     = '\033[32m'
-    YELLOW    = '\033[33m'
-    BLUE      = '\033[34m'
-    PURPLE    = '\033[35m'
     CYAN      = '\033[36m'
-    WHITE     = '\033[37m'
     END       = '\033[0m'
-    BOLD      = '\038[1m'
-    UNDERLINE = '\033[4m'
-    INVISIBLE = '\033[08m'
-    REVERCE   = '\033[07m'
+