+++ main/module/vvirt.py
@@ -1,8 +1,13 @@
-import setting
-import sys, sqlite3, subprocess, os
+
+import sys
+import sqlite3
+import subprocess
 import xml.etree.ElementTree as ET
-import libvirt, pings
-import vsql, vansible, vhelp,os, uuid
+import libvirt
+import pings
+import os
+import uuid
+from module import vsql, vansible, vhelp, setting